site fixes, sample updates, and designer package fixes
