+++ source/nodejs/adaptivecards-designer/cdn.html
@@ -4,9 +4,11 @@
 <head>
 	<title>Adaptive Cards Designer</title>
 	<script src="https://unpkg.com/markdown-it@8.4.0/dist/markdown-it.min.js"></script>
-	<script src="https://unpkg.com/adaptivecards-designer@0.1.1/dist/monaco-editor/min/vs/loader.js"></script>
 
-	<script src="https://unpkg.com/adaptivecards-designer@0.1.1/dist/adaptivecards-designer.min.js"></script>
+	<!-- Use your own copy of monaco if you prefer -->
+	<script src="https://unpkg.com/monaco-editor@0.17.1/min/vs/loader.js"></script>
+
+	<script src="https://unpkg.com/adaptivecards-designer@0.7.0/dist/adaptivecards-designer.min.js"></script>
 
 	<script type="text/javascript">
 		window.onload = function() {
@@ -19,11 +21,12 @@
 			hostContainers.push(new ACDesigner.CortanaContainer("Cortana Skills", "containers/cortana-container.css"));
 
 			let designer = new ACDesigner.CardDesigner(hostContainers);
-			designer.assetPath = "https://unpkg.com/adaptivecards-designer@0.1.1/dist";
+			designer.assetPath = "https://unpkg.com/adaptivecards-designer@0.7.0/dist";
 
 			designer.attachTo(document.getElementById("designerRootHost"));
 
-			require.config({ paths: { 'vs': 'https://unpkg.com/adaptivecards-designer@0.1.1/dist/monaco-editor/min/vs' } });
+			// initialize monaco editor and tell the Designer when it's been loaded
+			require.config({ paths: { 'vs': 'https://unpkg.com/monaco-editor@0.17.1/min/vs' } });
 			require(['vs/editor/editor.main'], function () {
 				designer.monacoModuleLoaded();
 			});
