+++ source/nodejs/adaptivecards-designer/index.html
@@ -4,7 +4,7 @@
 <head>
 	<title>Adaptive Cards Designer</title>
 	<script src="https://unpkg.com/markdown-it@8.4.0/dist/markdown-it.min.js"></script>
-	<script src="monaco-editor/min/vs/loader.js"></script>
+	<script src="https://unpkg.com/monaco-editor@0.17.1/min/vs/loader.js"></script>
 
 	<style type="text/css">
 		* {
@@ -84,7 +84,7 @@
 
 			require.config({
 				paths: {
-					'vs': 'monaco-editor/min/vs'
+					'vs': 'https://unpkg.com/monaco-editor@0.17.1/min/vs'
 				}
 			});
 			require(['vs/editor/editor.main'], function () {
