+++ source/nodejs/adaptivecards-designer/package.json
@@ -13,7 +13,8 @@
 	],
 	"scripts": {
 		"clean": "rimraf build lib dist",
-		"prebuild": "tsc",
+		"copy-hostconfigs": "cpy ../../../samples/HostConfig/*.json ./src/hostConfigs",
+		"prebuild": "npm run copy-hostconfigs && tsc",
 		"build": "webpack",
 		"watch": "webpack --watch",
 		"start": "webpack-dev-server --open",
@@ -21,18 +22,20 @@
 		"lint": "tslint -c tslint.json 'src/**/*.{ts,tsx}'",
 		"release": "npm run build && webpack --mode=production && npm run dts"
 	},
-	"dependencies": {
-		"adaptivecards": "^1.2.0",
-		"adaptivecards-controls": "^0.2.0",
-		"clipboard": "^2.0.1",
+	"peerDependencies": {
 		"monaco-editor": "^0.15.6"
 	},
 	"devDependencies": {
+		"adaptivecards": "^1.2.0",
+		"adaptivecards-controls": "^0.3.0",
+		"clipboard": "^2.0.1",
+		"cpy-cli": "^2.0.0",
+		"monaco-editor": "^0.17.1",
 		"rimraf": "^2.6.2",
 		"typescript": "^3.2.2",
-		"webpack": "^4.27.1",
-		"webpack-cli": "^3.1.2",
+		"webpack": "^4.35.0",
+		"webpack-cli": "^3.3.5",
 		"webpack-concat-plugin": "^3.0.0",
-		"webpack-dev-server": "^3.1.14"
+		"webpack-dev-server": "^3.7.2"
 	}
 }
