+++ source/nodejs/adaptivecards-designer/src/containers/default/adaptivecards-defaulthost.css
@@ -34,24 +34,32 @@
     cursor: pointer;
 }
 
+.ac-image.ac-selectable:hover {
+    background-color: rgba(0, 0, 0, 0.1);
+}
+
+.ac-image.ac-selectable:active {
+    background-color: rgba(0, 0, 0, 0.15);
+}
+
 a.ac-anchor {
     text-decoration: none;
 }
 
 a.ac-anchor:link {
-    color: #005A9E;
+    color: #005a9e;
 }
 
 a.ac-anchor:visited {
-    color: #005A9E;
+    color: #005a9e;
 }
 
 a.ac-anchor:link:active {
-    color: #004D84;
+    color: #004d84;
 }
 
 a.ac-anchor:visited:active {
-    color: #004D84;
+    color: #004d84;
 }
 
 .ac-container.ac-selectable, .ac-columnSet.ac-selectable {
@@ -67,72 +75,63 @@ a.ac-anchor:visited:active {
 }
 
 .ac-pushButton {
+    text-align: center;
+    vertical-align: middle;
+    cursor: default;
+    font-family: "Calibri", sans-serif;
+    font-size: 14px;
+    font-weight: 600;
+    padding: 8px;
+    -webkit-user-select: none;
+    -moz-user-select: none;
+    -ms-user-select: none;
+    user-select: none;
+    border: 1px solid #DDDDDD;
+		background-color: #FFFFFF;
+		color: #0063B1;
+}
+
+.ac-pushButton:focus {
+    border-color: #0063B1;
+}
+
+.ac-pushButton.style-positive {
+    background-color: #0078D7;
+    color: white;
+}
+
+.ac-pushButton.style-positive:hover, .ac-pushButton.style-positive:active {
+    background-color: #006ABC;
+}
+
+.ac-pushButton.style-destructive {
+    background-color: #E50000;
+    color: white;
+}
+
+.ac-pushButton.style-destructive:hover, .ac-pushButton.style-destructive:active {
+    background-color: #BF0000;
+}
+
+.ac-quickActionButton {
     overflow: hidden;
     text-overflow: ellipsis;
     text-align: center;
     vertical-align: middle;
     cursor: default;
-    font-family: "Segoe UI", sans-serif;
+    font-family: "Calibri", sans-serif;
     font-size: 14px;
     font-weight: 600;
-    padding: 4px 10px 5px 10px;
     -webkit-user-select: none;
     -moz-user-select: none;
     -ms-user-select: none;
     user-select: none;
-    height: 31px;
-    background-color: white;
-    color: #0078D7;
-    border: 1px solid #B2E0FF;
-}
-
-.ac-pushButton:hover {
-    background-color: #0078D7;
-    color: white;
-    border: 1px solid #73C1FD;
-}
-
-.ac-pushButton:active {
-    background-color: #464B93;
-    color: white;
-    border: 1px solid #464B93;
-}
-
-.ac-pushButton.subdued {
-    color: #666666;
-    border: 1px solid #EEEEEE;
-}
-
-.ac-pushButton.subdued:hover {
-    background-color: #0078D7;
-    color: white;
-    border: 1px solid #0078D7;
-}
-
-.ac-pushButton.expanded {
-    background-color: #0078D7;
-    color: white;
-    border: 1px solid #0078D7;
-}
-
-.ac-pushButton.expandable:after {
-    -moz-osx-font-smoothing: grayscale;
-    -webkit-font-smoothing: antialiased;
-    display: inline-block;
-    font-family: 'FabricMDL2Icons';
-    font-size: 10px;
-    font-style: normal;
-    font-weight: 900;
-    margin: auto 0px auto 6px;
-    content: '\E70D';
-}
-
-.ac-pushButton.expandable.expanded:after {
-    content: '\E70E';
+    border: none;
+    background-color: #DDDDDD;
 }
 
 .ac-input {
-    font-family: "Segoe UI", sans-serif;
+    font-family: "Calibri", sans-serif;
     font-size: 14px;
     color: black;
 }
@@ -148,6 +147,11 @@ a.ac-anchor:visited:active {
     border: 1px solid red !important;
 }
 
+.ac-input-container {
+    display: flex;
+    align-items: flex-end;
+}
+
 .ac-input.ac-textInput {
     resize: none;
 }
@@ -161,7 +165,7 @@ a.ac-anchor:visited:active {
 }
 
 .ac-input.ac-textInput, .ac-input.ac-numberInput, .ac-input.ac-dateInput, .ac-input.ac-timeInput, .ac-input.ac-multichoiceInput {
-    border: 1px solid #DDDDDD;
+    border: 1px solid #dddddd;
     padding: 4px 8px 4px 8px;
 }
 
@@ -173,8 +177,8 @@ a.ac-anchor:visited:active {
     text-align: center;
     vertical-align: middle;
     cursor: pointer;
-    font-family: "Segoe UI", sans-serif;
-    font-size: 14px;
+    font-family: "Calibri", sans-serif;
+    font-size: 13.3px;
     -webkit-user-select: none;
     -moz-user-select: none;
     -ms-user-select: none;
