+++ source/nodejs/adaptivecards-designer/src/containers/notifications/toast-container.ts
@@ -2,8 +2,7 @@
 // Licensed under the MIT License.
 import * as Adaptive from "adaptivecards";
 import * as Designer from "../../adaptivecards-designer";
-
-var windowsNotificationConfiguration = require("../../../../../../samples/HostConfig/windows-notification.json");
+import * as hostConfig from "../../hostConfigs/windows-notification.json";
 
 export class ToastContainer extends Designer.HostContainer {
     public renderTo(hostElement: HTMLElement) {
@@ -17,6 +16,6 @@ export class ToastContainer extends Designer.HostContainer {
     }
 
     public getHostConfig(): Adaptive.HostConfig {
-        return new Adaptive.HostConfig(windowsNotificationConfiguration);
+        return new Adaptive.HostConfig(hostConfig);
     }
 }
