+++ source/nodejs/adaptivecards-designer/src/containers/teams/teams-container.ts
@@ -2,9 +2,8 @@
 // Licensed under the MIT License.
 import * as Adaptive from "adaptivecards";
 import * as Designer from "../../adaptivecards-designer";
-
-var teamsDarkConfiguration = require("../../../../../../samples/HostConfig/microsoft-teams-dark.json");
-var teamsLightConfiguration = require("../../../../../../samples/HostConfig/microsoft-teams-light.json");
+import * as hostConfigLight from "../../hostConfigs/microsoft-teams-light.json";
+import * as hostConfigDark from "../../hostConfigs/microsoft-teams-dark.json";
 
 export abstract class BaseTeamsContainer extends Designer.HostContainer {
     public renderTo(hostElement: HTMLElement) {
@@ -46,7 +45,7 @@ export abstract class BaseTeamsContainer extends Designer.HostContainer {
 
 export class LightTeamsContainer extends BaseTeamsContainer {
     public getHostConfig(): Adaptive.HostConfig {
-        return new Adaptive.HostConfig(teamsLightConfiguration);
+        return new Adaptive.HostConfig(hostConfigLight);
     }
 }
 
@@ -56,6 +55,6 @@ export class DarkTeamsContainer extends BaseTeamsContainer {
     }
 
     public getHostConfig(): Adaptive.HostConfig {
-        return new Adaptive.HostConfig(teamsDarkConfiguration);
+        return new Adaptive.HostConfig(hostConfigDark);
     }
 }
