+++ source/nodejs/adaptivecards-designer/src/hostConfigs/README.md
@@ -1,0 +1,5 @@
+**IMPORTANT**
+
+These Host Configs are pulled from the `~/samples/HostConfig` directory and **SHOULD NOT** be modified here or committed. 
+
+They will be overwritten on build.
