+++ source/nodejs/adaptivecards-designer/tsconfig.json
@@ -1,7 +1,8 @@
 {
 	"extends": "../tsconfig.json",
 	"compilerOptions": {
-        "outDir": "./lib"
+		"outDir": "./lib",
+		"resolveJsonModule": true
 	},
-	"files": ["./src/adaptivecards-designer.ts", "./src/containers/index.ts"]
+	"files": ["./src/adaptivecards-designer.ts" ]
 }
