+++ source/nodejs/adaptivecards-site/package.json
