+++ source/nodejs/adaptivecards-site/themes/adaptivecards/layout/_partial/head.ejs
@@ -83,9 +83,7 @@
 		<%- js('node_modules/monaco-editor/min/vs/loader.js') %>  
 		<%- js("node_modules/adaptivecards-designer/dist/adaptivecards-designer.min.js") %>
 	<% } else { %>
-		<%- js('node_modules/adaptivecards/dist/adaptivecards.min.js') %>
-	
-		<%- js('js/stickyfill.min.js') %>
+		<%- js('node_modules/adaptivecards/dist/adaptivecards.min.js') %>	
 		<%- js('js/script.js') %>
 	<% } %>
