+++ source/nodejs/adaptivecards-site/themes/adaptivecards/package.json
