+++ source/nodejs/tsconfig.json
@@ -9,7 +9,7 @@
 		"moduleResolution": "node",
 		"noImplicitAny": false,
 		"sourceMap": true,
-		"declaration": true,
+		"declaration": true
 	},
 	"exclude": [
 		"node_modules",
